+++ src/pytest_given/__init__.py
@@ -14,12 +14,13 @@ from .capture import (
     when,
     when_then,
 )
-from .model import PytestGivenError
+from .model import PytestGivenError, PytestGivenWarning
 
 __all__ = [
     'FileGlossary',
     'Glossary',
     'PytestGivenError',
+    'PytestGivenWarning',
     'Template',
     'activity',
     'attach',